backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
sqlalchemy>=2.0.27
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
python-dotenv>=1.0.1
mediapipe>=0.10.9
opencv-python>=4.9.0.80
numpy>=1.24.3
scikit-learn
requests
facenet-pytorch
torchvision
email-validator
